pyproject.toml
[project]
name = "image-tools"
version = "0.1.0"
description = "A collection of simple image tools for renaming and converting images."
authors = [{ name = "Guillaume Mousnier", email = "guillaume.mousnier@datawizard.fr" }]

dependencies = [
    "Pillow>=10.0.0",
    "pyheif>=0.7.0"
]

[project.scripts]
convert-heic-to-jpg = "image_tools.convert_heic_to_jpg:main"
rename-files-by-date = "image_tools.rename_files_by_date:main"

[dependency-groups]
dev = ["pytest", "flake8", "black"]